some refactoring and start writing the inserting_info done with inserting players and commanders. need to update commander elo the insertion is still inefficient.
